before fixing serious lock problem

+++ autouri/gcsuri.py
@@ -1,43 +1,22 @@
-#!/usr/bin/env python3
-"""
-Bucket rules:
+"""GCS Bucket rules:
     Object versioning must be turned off
         https://cloud.google.com/storage/docs/object-versioning
 """
 import os
 import requests
-import binascii
-from base64 import b64decode
-from contextlib import contextmanager
-from datetime import (datetime, timedelta)
-from dateutil.parser import parse as parse_timestamp
-from dateutil.tz import tzutc
+import time
 from filelock import BaseFileLock
 from google.api_core.exceptions import NotFound, Forbidden, GatewayTimeout, ServiceUnavailable
-# from requests.exceptions import ConnectionError
 from google.cloud import storage
 from google.cloud.storage import Blob
 from google.oauth2.service_account import Credentials
-from subprocess import (
-    check_output, check_call, Popen, PIPE, CalledProcessError)
+from subprocess import check_call, PIPE, CalledProcessError
+from tempfile import NamedTemporaryFile
 from typing import Tuple, Optional
-from .autouri import URIBase, URIMetadata, AutoURI, logger
+from urllib3.exceptions import HTTPError
 
-
-def init_gcsuri(
-    loc_prefix: Optional[str]=None,
-    private_key_file: Optional[str]=None,
-    sec_duration_presigned_url: Optional[int]=None):
-    """Helper function to initialize GCSURI class constants
-        loc_prefix:
-            Inherited from URIBase
-    """
-    if loc_prefix is not None:
-        GCSURI.LOC_PREFIX = loc_prefix
-    if private_key_file is not None:
-        GCSURI.PRIVATE_KEY_FILE = private_key_file
-    if sec_duration_presigned_url is not None:
-        GCSURI.SEC_DURATION_PRESIGNED_URL = sec_duration_presigned_url
+from .autouri import URIBase, AutoURI, logger
+from .metadata import URIMetadata, get_seconds_from_epoch, parse_md5_str
 
 
 class GCSURILock(BaseFileLock):
@@ -53,7 +32,7 @@ class GCSURILock(BaseFileLock):
 
     def _acquire(self):
         u = GCSURI(self._lock_file)
-        blob = u.get_blob(new=True)
+        blob, _ = u.get_blob(new=True)
         if blob is not None:
             try:
                 blob.upload_from_string('')
@@ -66,27 +45,32 @@ class GCSURILock(BaseFileLock):
 
     def _release(self):
         u = GCSURI(self._lock_file)
-        blob = u.get_blob()
-        blob.temporary_hold = False
-        try:
-            # u.rm(no_lock=True)
-            blob.patch()
-            blob.delete()
-            self._lock_file_fd = None
-        except (NotFound,):
-            pass
+        blob, _ = u.get_blob()
+        if blob is not None:
+            blob.temporary_hold = False
+            try:
+                blob.patch()
+                blob.delete()
+                self._lock_file_fd = None
+            except (NotFound,):
+                pass
         return None
 
 
 class GCSURI(URIBase):
     """
     Class constants:
-        LOC_PREFIX:
+        LOC_PREFIX (inherited):
             Path prefix for localization. Inherited from URIBase class.
         PRIVATE_KEY_FILE:
             Path for private key file used to get presigned URLs
-        SEC_DURATION_PRESIGNED_URL:
+        DURATION_PRESIGNED_URL:
             Duration for presigned URLs
+        RETRY_BUCKET:
+            Number of retrial to access a bucket
+        RETRY_BUCKET_DELAY:
+            Delay for each retrial in seconds
+
     Protected class constants:
         _CACHED_GCS_CLIENT_PER_THREAD:
             Per-thread GCS client object is required since 
@@ -94,8 +78,11 @@ class GCSURI(URIBase):
         _CACHED_PRESIGNED_URLS:
             Can use cached presigned URLs.
     """
-    PRIVATE_KEY_FILE: str = None
-    SEC_DURATION_PRESIGNED_URL: int = 4233600
+    PRIVATE_KEY_FILE: str = ''
+    DURATION_PRESIGNED_URL: int = 4233600
+
+    RETRY_BUCKET: int = 3
+    RETRY_BUCKET_DELAY: int = 1
 
     _CACHED_GCS_CLIENT_PER_THREAD = {}
     _CACHED_PRESIGNED_URLS = {}
@@ -106,15 +93,13 @@ class GCSURI(URIBase):
     def __init__(self, uri, thread_id=-1):
         super().__init__(uri, thread_id=thread_id)
 
-    def get_lock(self, no_lock=False, timeout=None, poll_interval=None):
-        if no_lock:
-            return contextmanager(lambda: (yield))()
+    def _get_lock(self, timeout=None, poll_interval=None):
         if timeout is None:
             timeout = GCSURI.LOCK_TIMEOUT
         if poll_interval is None:
             poll_interval = GCSURI.LOCK_POLL_INTERVAL
         return GCSURILock(
-            self._uri + AutoURI.LOCK_FILE_EXT,
+            self._uri + GCSURI.LOCK_FILE_EXT,
             timeout=timeout,
             poll_interval=poll_interval)
 
@@ -122,41 +107,28 @@ class GCSURI(URIBase):
         ex, mt, sz, md5 = False, None, None, None
 
         try:
-            b = self.get_blob()
+            b, _ = self.get_blob()
             if b is not None:
                 # make keys lower-case
                 h = {k.lower(): v for k, v in b._properties.items()}
                 ex = True
 
-                md5_raw = None
-                if 'md5hash' in h:
-                    md5_raw = h['md5hash']
-                elif 'etag' in h:
-                    md5_raw = h['etag']
-                if md5_raw is not None:
-                    md5_raw = md5_raw.strip('"\'')
-                    if len(md5_raw) == 32:
-                        md5 = md5_raw
-                    else:
-                        md5 = binascii.hexlify(b64decode(md5_raw)).decode()
+                if not skip_md5:
+                    if 'md5hash' in h:
+                        md5 = parse_md5_str(h['md5hash'])
+                    elif 'etag' in h:
+                        md5 = parse_md5_str(h['etag'])
+                    if md5 is None:
+                        # make_md5_file is ignored for GCSURI
+                        md5 = self.md5_from_file
 
                 if 'size' in h:
                     sz = int(h['size'])
 
                 if 'updated' in h:
-                    utc_t = parse_timestamp(h['updated'])
+                    mt = get_seconds_from_epoch(h['updated'])
                 elif 'timecreated' in h:
-                    utc_t = parse_timestamp(h['timecreated'])
-                else:
-                    utc_t = None
-                if utc_t is not None:              
-                    utc_epoch = datetime(1970, 1, 1, tzinfo=tzutc())      
-                    mt = (utc_t - utc_epoch).total_seconds()
-
-                if md5 is None and not skip_md5:
-                    md5 = self.md5_from_file
-                    # make_md5_file is ignored for GCSURI
-                        
+                    mt = get_seconds_from_epoch(h['timecreated'])
 
         except Exception as e:
             pass
@@ -168,19 +140,20 @@ class GCSURI(URIBase):
             md5=md5)
 
     def read(self, byte=False):
-        blob = self.get_blob()
+        blob, _ = self.get_blob()
         b = blob.download_as_string()
         if byte:
             return b
         return b.decode()
 
     def _write(self, s):
-        blob = self.get_blob(new=True)
+        blob, _ = self.get_blob(new=True)
         blob.upload_from_string(s)
+        # blob.update()
         return
 
     def _rm(self):
-        blob = self.get_blob()
+        blob, _ = self.get_blob()
         blob.delete()
         return
 
@@ -196,15 +169,17 @@ class GCSURI(URIBase):
         dest_uri = AutoURI(dest_uri)
 
         if isinstance(dest_uri, (GCSURI, AbsPath)):            
-            src_blob = self.get_blob()
+            src_blob, src_bucket = self.get_blob()
 
             if isinstance(dest_uri, GCSURI):
-                dest_bucket, dest_path = dest_uri.get_bucket_path()
+                _, dest_path = dest_uri.get_bucket_path()
+                _, dest_bucket = dest_uri.get_blob()
                 src_bucket.copy_blob(src_blob, dest_bucket, dest_path)
                 return True
 
             elif isinstance(dest_uri, AbsPath):
                 dest_uri.mkdir_dirname()
+                src_blob.update()
                 src_blob.download_to_filename(dest_uri._uri)
                 return True
 
@@ -221,11 +196,12 @@ class GCSURI(URIBase):
         """
         from .s3uri import S3URI
         from .abspath import AbsPath
+        from .httpurl import HTTPURL
 
         src_uri = AutoURI(src_uri)
 
         if isinstance(src_uri, AbsPath):
-            blob = self.get_blob(new=True)
+            blob, _ = self.get_blob(new=True)
             blob.upload_from_filename(src_uri._uri)
             return True
 
@@ -238,29 +214,43 @@ class GCSURI(URIBase):
                 src_uri._uri, stream=True, allow_redirects=True,
                 headers=requests.utils.default_headers())
             r.raise_for_status()
-            b = BytesIO()
-            with open(b, 'wb') as fp:
-                for chunk in r.iter_content(HTTPURL.get_chunk_size()):
+            with NamedTemporaryFile() as fp:
+                for chunk in r.iter_content(HTTPURL.get_http_chunk_size()):
                     fp.write(chunk)
-            blob = self.get_blob(new=True)
-            with open(b, 'rb') as fp:
+                fp.seek(0)
+                blob, _ = self.get_blob(new=True)
                 blob.upload_from_file(fp)
             return True
         return False
 
     def get_blob(self, new=False) -> Blob:
-        """GCS client has a bug that shows an outdated version of a file
+        """GCS client() has a bug that shows an outdated version of a file
         when using Blob() without update().
-
         For read-only functions (e.g. read()), need to directly call
         cl.get_bucket(bucket).get_blob(path) instead of using Blob() class.
+
+        Also, GCS client() is not thread-safe and it fails for a variety of reasons.
+        Retry several times for whatever reasons.
+
+        Returns:
+            blob: Blob object or None
+            bucket_obj: Bucket object or None
         """
         bucket, path = self.get_bucket_path()
         cl = GCSURI.get_gcs_client(self._thread_id)
-        if new:
-            return Blob(name=path, bucket=cl.get_bucket(bucket))
-        else:
-            return cl.get_bucket(bucket).get_blob(path)
+
+        bucket_obj = None
+        blob = None
+        for retry in range(GCSURI.RETRY_BUCKET):
+            try:
+                bucket_obj = cl.get_bucket(bucket)
+                blob = bucket_obj.get_blob(path)
+                if new and blob is None:
+                    blob = Blob(name=path, bucket=bucket_obj)
+                break
+            except (Exception,):
+                time.sleep(GCSURI.RETRY_BUCKET_DELAY)
+        return blob, bucket_obj
 
     def get_bucket_path(self) -> Tuple[str, str]:
         """Returns a tuple of URI's S3 bucket and path.
@@ -268,10 +258,10 @@ class GCSURI(URIBase):
         bucket, path = self.uri_wo_scheme.split(GCSURI.get_path_sep(), 1)
         return bucket, path        
 
-    def get_presigned_url(self, sec_duration=None, use_cached=False) -> str:
+    def get_presigned_url(self, duration=None, use_cached=False) -> str:
         """
         Args:
-            sec_duration: Duration in seconds. This is ignored if use_cached is on.
+            duration: Duration in seconds. This is ignored if use_cached is on.
             use_cached: Use a cached URL. 
         """
         cache = GCSSURI._CACHED_PRESIGNED_URLS
@@ -281,9 +271,10 @@ class GCSURI(URIBase):
         blob = self.get_blob()
         private_key_file = os.path.expanduser(GCSURI.PRIVATE_KEY_FILE)
         if not os.path.exists(private_key_file):
-            raise Exception('GCS private key file not found')
+            raise Exception('GCS private key file not found. f:{f}'.format(
+                f=private_key_file))
         credentials = Credentials.from_service_account_file(private_key_file)
-        duration = sec_duration if sec_duration is not None else GSSURI.SEC_DURATION_PRESIGNED_URL        
+        duration = duration if duration is not None else GSSURI.DURATION_PRESIGNED_URL        
         url = blob.generate_signed_url(
             expiration=timedelta(seconds=duration),
             credentials=credentials)
@@ -298,3 +289,21 @@ class GCSURI(URIBase):
             cl = storage.Client()
             GCSURI._CACHED_GCS_CLIENT_PER_THREAD[thread_id] = cl
             return cl
+
+    @staticmethod
+    def init_gcsuri(
+        loc_prefix: Optional[str]=None,
+        private_key_file: Optional[str]=None,
+        duration_presigned_url: Optional[int]=None,
+        retry_bucket: Optional[int]=None,
+        retry_bucket_delay: Optional[int]=None):
+        if loc_prefix is not None:
+            GCSURI.LOC_PREFIX = loc_prefix
+        if private_key_file is not None:
+            GCSURI.PRIVATE_KEY_FILE = private_key_file
+        if duration_presigned_url is not None:
+            GCSURI.DURATION_PRESIGNED_URL = duration_presigned_url
+        if retry_bucket is not None:
+            GCSURI.RETRY_BUCKET = retry_bucket
+        if retry_bucket_delay is not None:
+            GCSURI.RETRY_BUCKET_DELAY = retry_bucket_delay
